requirements.txt
flask>=3.0
tensorflow==2.20.0
pillow>=10.0
numpy>=1.25
h5py>=3.10
